+++ tasks/api/src/main/java/org/projectnessie/nessie/tasks/api/TaskState.java
@@ -69,27 +69,50 @@ public interface TaskState {
   @JsonInclude(JsonInclude.Include.NON_NULL)
   String message();
 
-  TaskState SUCCESS = ImmutableTaskState.of(TaskStatus.SUCCESS, null, null, null);
+  /**
+   * An error code associated with this task state. Error codes are machine-readable representations
+   * of failures and are meant to be interpreted by the {@link TaskBehavior} implementation that
+   * manages the related task.
+   *
+   * <p>On the other hand, the value of {@link #message()} is meant to be interpreted by humans.
+   *
+   * <p>Storing error codes in {@link TaskState} is intended to allow reporting task errors /
+   * exceptions to the caller the same way whether the task fails for the first time in the server
+   * that executed it, or it is reloaded from storage in another server.
+   *
+   * @see TaskBehavior#stateAsException(TaskObj)
+   */
+  @Value.Parameter(order = 5)
+  @Nullable
+  @JsonInclude(JsonInclude.Include.NON_NULL)
+  String errorCode();
+
+  TaskState SUCCESS = ImmutableTaskState.of(TaskStatus.SUCCESS, null, null, null, null);
 
   static TaskState successState() {
     return SUCCESS;
   }
 
   static TaskState runningState(@Nonnull Instant retryNotBefore, @Nonnull Instant lostNotBefore) {
-    return ImmutableTaskState.of(TaskStatus.RUNNING, retryNotBefore, lostNotBefore, null);
+    return ImmutableTaskState.of(TaskStatus.RUNNING, retryNotBefore, lostNotBefore, null, null);
   }
 
-  static TaskState retryableErrorState(@Nonnull Instant retryNotBefore, @Nonnull String message) {
-    return ImmutableTaskState.of(TaskStatus.ERROR_RETRY, retryNotBefore, null, message);
+  static TaskState retryableErrorState(
+      @Nonnull Instant retryNotBefore, @Nonnull String message, @Nonnull String errorCode) {
+    return ImmutableTaskState.of(TaskStatus.ERROR_RETRY, retryNotBefore, null, message, errorCode);
   }
 
-  static TaskState failureState(@Nonnull String message) {
-    return ImmutableTaskState.of(TaskStatus.FAILURE, null, null, message);
+  static TaskState failureState(@Nonnull String message, String errorCode) {
+    return ImmutableTaskState.of(TaskStatus.FAILURE, null, null, message, errorCode);
   }
 
   static TaskState taskState(
-      TaskStatus taskStatus, Instant retryNotBefore, Instant lostNotBefore, String message) {
-    return ImmutableTaskState.of(taskStatus, retryNotBefore, lostNotBefore, message);
+      TaskStatus taskStatus,
+      Instant retryNotBefore,
+      Instant lostNotBefore,
+      String message,
+      String errorCode) {
+    return ImmutableTaskState.of(taskStatus, retryNotBefore, lostNotBefore, message, errorCode);
   }
 
   @Value.Check